split off config_keeperfx
