+++ src/moonphase.c
@@ -3,6 +3,12 @@
 #include <astronomy.h>
 #include "post_inc.h"
 
+short is_full_moon = 0;
+short is_near_full_moon = 0;
+short is_new_moon = 0;
+short is_near_new_moon = 0;
+
+
 /**
  * Calculate the moon phase.
  *
@@ -13,7 +19,7 @@
  *
  * Returns -1 on failure
  */
-double moonphase_calculate()
+static double moonphase_calculate()
 {
     // Get current time
     astro_time_t time = Astronomy_CurrentTime();
@@ -26,4 +32,81 @@ double moonphase_calculate()
     }
 
     return phase.angle / 360;
+}
+
+short calculate_moon_phase(short do_calculate, short add_to_log)
+{
+    // Moon phase calculation
+    if (do_calculate)
+    {
+        phase_of_moon = (float)moonphase_calculate();
+    }
+
+    // Handle moon phases
+    if ((phase_of_moon > 0.475) && (phase_of_moon < 0.525)) // Approx 33 hours
+    {
+        if (add_to_log)
+        {
+            SYNCMSG("Full moon %.4f", phase_of_moon);
+        }
+
+        is_full_moon = 1;
+        is_near_full_moon = 0;
+        is_new_moon = 0;
+        is_near_new_moon = 0;
+    }
+    else if ((phase_of_moon > 0.45) && (phase_of_moon < 0.55)) // Approx 70 hours
+    {
+        if (add_to_log)
+        {
+            SYNCMSG("Near Full moon %.4f", phase_of_moon);
+        }
+
+        is_full_moon = 0;
+        is_near_full_moon = 1;
+        is_new_moon = 0;
+        is_near_new_moon = 0;
+    }
+    else if ((phase_of_moon < 0.025) || (phase_of_moon > 0.975))
+    {
+        if (add_to_log)
+        {
+            SYNCMSG("New moon %.4f", phase_of_moon);
+        }
+
+        is_full_moon = 0;
+        is_near_full_moon = 0;
+        is_new_moon = 1;
+        is_near_new_moon = 0;
+    }
+    else if ((phase_of_moon < 0.05) || (phase_of_moon > 0.95))
+    {
+        if (add_to_log)
+        {
+            SYNCMSG("Near new moon %.4f", phase_of_moon);
+        }
+
+        is_full_moon = 0;
+        is_near_full_moon = 0;
+        is_new_moon = 0;
+        is_near_new_moon = 1;
+    }
+    else
+    {
+        if (add_to_log)
+        {
+            SYNCMSG("Moon phase %.4f", phase_of_moon);
+        }
+
+        is_full_moon = 0;
+        is_near_full_moon = 0;
+        is_new_moon = 0;
+        is_near_new_moon = 0;
+    }
+
+    //! CHEAT! always show extra levels
+    // TODO: make this a command line option?
+    //  is_full_moon = 1; is_new_moon = 1;
+
+    return is_full_moon;
 }
