+++ src/moonphase.h
@@ -6,7 +6,12 @@ extern "C"
 {
 #endif
 
-    double moonphase_calculate();
+    extern short is_full_moon;
+    extern short is_near_full_moon;
+    extern short is_new_moon;
+    extern short is_near_new_moon;
+    
+    short calculate_moon_phase(short do_calculate, short add_to_log)
 
 #ifdef __cplusplus
 }
